fix not found page error
