+++ components/ui/toast.tsx
@@ -12,6 +12,7 @@ export const Toaster = ({ ...props }: ToasterProps) => {
     <Sonner
       theme={theme as ToasterProps['theme']}
       className='toaster group'
+      richColors
       toastOptions={{
         classNames: {
           toast:
